refactor(DB): 도메인별로 패키지 분리
