+++ src/main/java/roomescape/reservation/ReservationController.java
@@ -1,4 +1,4 @@
-package roomescape;
+package roomescape.reservation;
 
 import java.util.List;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -8,7 +8,6 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.domain.Reservation;
 
 @RestController
 @RequestMapping("/reservations")
